Only enable JRuby objectspace in tests
This is not ideal because it's possible for features to be introduced
that rely on objectspace which will break in JRuby
but pass in the tests. Ideally, the tests would eventually
be refactored to not rely on objectspace.

We don't want objectspace in runtime because it decreases performance.
From the JRuby wiki:

> ObjectSpace has been disabled by default since version 1.1b1. Some users reenable ObjectSpace, using the -X+O flag (previously +O), without realizing what a massive performance hit this is. If you are running with ObjectSpace enabled, do not expect any sort of performance.

https://github.com/jruby/jruby/wiki/PerformanceTuning#dont-enable-objectspace