Add an option to create temp files in the zip file directory
Before zip files were created in the system temporary directory they
were created in the same directory as the zip file. This adds an option
to have the same behaviour as before by passing in
`use_system_temp_dir: false` as an option to `Zip.open`, which passes
that down to `Zip::StreamableStream`.