Update some dev dependency gems and relax version matching.
For gems such as rake, minitest and rdoc it is appropriate to allow version
matches against major versions, allowing us to get the latest minor and
patch versions automatically.

Also update the minimum minor versions while we are at it.

We'll leave rubocop* and simplecov* gems pinned to minor versions.