Eagerly require gem version
v3 is released and contains breaking changes. I was a bit surprised that I had to manually require this file.

Can't be helped now of course but it will help with future major versions.

Nodoc on it also seems wrong. To support both v2 and v3, I have to know which version I'm using.