Fix reading EOCDs when header signatures are in an Entry payload.
We were naively searching for the Zip64 EOCD data structures in a Zip,
even if they weren't needed, which caused an issue if they weren't there
but an Entry payload contained the header signature.

Now we only read the Zip64 EOCDs if we know we need them to complete
our model of the Zip archive. This aught to be robust against similar
situations in future too.

Fixes #656